remove 4x and move player movement out of input
