+++ senior_game_client/Cargo.toml
@@ -14,7 +14,6 @@ bevy_asset_ron = "0.1.0"
 bevy-inspector-egui = "0.3.0"
 bevy_egui = "0.3.0"
 kurinji = "1.0.5"
-bevy_4x_camera = "0.1.2"
 bevy_mod_picking = "0.3.7"
 serde = "1.0.124"
 sentry = "0.22.0"
