+++ senior_game_shared/Cargo.toml
@@ -7,7 +7,6 @@ edition = "2018"
 [dependencies]
 bevy = "0.4.0"
 bevy_rapier3d = "0.8.0"
-bevy_4x_camera = "0.1.2"
 bevy_prototype_networking_laminar = { git = "https://github.com/Hugheth/bevy_prototype_networking_laminar", branch = "bevy-0.4" }
 serde = "1.0.124"
 bincode = "1.3.2"
